Rewrite notification, rearrange files, minor UI adjustment
