+++ main/footer.php
@@ -1,4 +1,14 @@
 <section class="footer fix">
+   <style>
+      #contact-email {
+         display:flex;
+         justify-content:center;
+      }
+      #contact-email > a {
+         padding:0;
+         margin:0;
+      }
+   </style>
    <div class="box-container">
       <div class="box">
          <h3>quick links</h3><br>
@@ -19,19 +29,19 @@
       <div class="box">
          <h3>contact info</h3><br>
          <p> <i class="fas fa-phone"></i> +123-456-7890 </p>
-         <p style="display:flex; justify-content:center; align-item:center;"><i class="fas fa-envelope"></i><a style="padding:0;margin:0;" href="mailto:trung150502@gmail.com">trung150502@gmail.com</a></p>
+         <p id="contact-email">
+            <i class="fas fa-envelope"></i>
+            <a href="mailto:trung150502@gmail.com">trung150502@gmail.com</a>
+         </p>
          <p> <i class="fas fa-map-marker-alt"></i>Ho Chi Minh City, VietNam</p>
       </div>
 
       <div class="box">
          <h3>follow us</h3><br>
-         <a href="#"> <i class="fab fa-facebook-f"></i> FACEBOOK </a>
-         <a href="#"> <i class="fab fa-twitter"></i> TWITTER </a>
-         <a href="#"> <i class="fab fa-instagram"></i> INSTAGRAM </a>
-         <a href="#"> <i class="fab fa-linkedin"></i> LINKEDIN </a>
+         <a href="#"> <i class="fab fa-facebook-f"></i>FACEBOOK</a>
+         <a href="#"> <i class="fab fa-twitter"></i>TWITTER</a>
+         <a href="#"> <i class="fab fa-instagram"></i>INSTAGRAM</a>
       </div>
-
    </div>
    <p class="credit">WEBSITE BY <span>Nhóm 3-lập trình Web-học kì 221-HCMUT</span> </p>
-   <p style="text-align:center;padding:2rem;">------------------------<p>
 </section>
